:truck: api: move get jury certification controller to src
