+++ api/lib/application/certification-courses/certification-course-controller.js
@@ -4,15 +4,6 @@ import { DomainTransaction } from '../../infrastructure/DomainTransaction.js';
 import * as certifiedProfileRepository from '../../infrastructure/repositories/certified-profile-repository.js';
 import * as certificationCourseSerializer from '../../infrastructure/serializers/jsonapi/certification-course-serializer.js';
 import * as certifiedProfileSerializer from '../../infrastructure/serializers/jsonapi/certified-profile-serializer.js';
-import * as juryCertificationSerializer from '../../infrastructure/serializers/jsonapi/jury-certification-serializer.js';
-
-const getJuryCertification = async function (request, h, dependencies = { juryCertificationSerializer }) {
-  const certificationCourseId = request.params.id;
-  const translate = request.i18n.__;
-  const juryCertification = await usecases.getJuryCertification({ certificationCourseId });
-
-  return dependencies.juryCertificationSerializer.serialize(juryCertification, { translate });
-};
 
 const save = async function (request, h, dependencies = { extractLocaleFromRequest, certificationCourseSerializer }) {
   const userId = request.auth.credentials.userId;
@@ -52,7 +43,6 @@ const getCertifiedProfile = async function (
 };
 
 const certificationCourseController = {
-  getJuryCertification,
   save,
   get,
   getCertifiedProfile,
